Move bundles_to_extrinsics to domain_block_processor
