+++ domains/client/domain-executor/src/domain_block_processor.rs
@@ -3,7 +3,7 @@ use crate::utils::{shuffle_extrinsics, to_number_primitive, translate_number_typ
 use crate::{ExecutionReceiptFor, TransactionFor};
 use codec::{Decode, Encode};
 use domain_block_builder::{BlockBuilder, BuiltBlock, RecordProof};
-use domain_runtime_primitives::{AccountId, DomainCoreApi};
+use domain_runtime_primitives::{AccountId, DomainCoreApi, DomainExtrinsicApi};
 use sc_client_api::{AuxStore, BlockBackend, StateBackendFor};
 use sc_consensus::{
     BlockImport, BlockImportParams, ForkChoiceStrategy, ImportResult, StateAction, StorageChanges,
@@ -164,6 +164,7 @@ where
     Client:
         HeaderBackend<Block> + BlockBackend<Block> + AuxStore + ProvideRuntimeApi<Block> + 'static,
     Client::Api: DomainCoreApi<Block, AccountId>
+        + DomainExtrinsicApi<Block, NumberFor<PBlock>, PBlock::Hash>
         + sp_block_builder::BlockBuilder<Block>
         + sp_api::ApiExt<Block, StateBackend = StateBackendFor<Backend, Block>>,
     for<'b> &'b Client: BlockImport<
@@ -318,6 +319,52 @@ where
                 }
             }
         }
+    }
+
+    pub(crate) fn bundles_to_extrinsics(
+        &self,
+        parent_hash: Block::Hash,
+        bundles: DomainBundles<Block, PBlock>,
+        shuffling_seed: Randomness,
+    ) -> Result<Vec<Block::Extrinsic>, sp_blockchain::Error> {
+        let extrinsics = match bundles {
+            DomainBundles::System(system_bundles, core_bundles) => {
+                if self.domain_id.is_core() {
+                    return Err(sp_blockchain::Error::Application(Box::from(
+                        "Core bundle processor can not process system bundles.",
+                    )));
+                }
+                let origin_system_extrinsics = self.compile_own_domain_bundles(system_bundles);
+                self
+                    .client
+                    .runtime_api()
+                    .construct_submit_core_bundle_extrinsics(&BlockId::Hash(parent_hash), core_bundles)?
+                    .into_iter()
+                    .filter_map(
+                        |uxt| match <<Block as BlockT>::Extrinsic>::decode(&mut uxt.as_slice()) {
+                            Ok(uxt) => Some(uxt),
+                            Err(e) => {
+                                tracing::error!(
+                                    error = ?e,
+                                    "Failed to decode the opaque extrisic in bundle, this should not happen"
+                                );
+                                None
+                            }
+                        },
+                    )
+                    .chain(origin_system_extrinsics)
+                    .collect::<Vec<_>>()
+            }
+            DomainBundles::Core(bundles) => {
+                if self.domain_id.is_system() {
+                    return Err(sp_blockchain::Error::Application(Box::from(
+                        "System bundle processor can not process core bundles.",
+                    )));
+                }
+                self.compile_own_domain_bundles(bundles)
+            }
+        };
+        self.deduplicate_and_shuffle_extrinsics(parent_hash, extrinsics, shuffling_seed)
     }
 
     pub(crate) fn compile_own_domain_bundles(
